tests work for postgres and mysql
